requirements.txt
streamlit
google-generativeai
pdfplumber
python-dotenv
